requirements.txt
Django==5.0.7
Markdown==3.6
python-dotenv==1.0.1
pymysql==1.1.1
line-bot-sdk==3.11.0